Rename "PendingOfferTab" => "PendingOffersTab", refactor sortViewParams, improve ts for optionKeys in searchSelect
