+++ src/components/SearchSelect/SearchSelect.tsx
@@ -12,7 +12,7 @@ import styles from './SearchSelect.module.less'
 
 export interface SearchSelectProps<P> {
   options: P[]
-  optionKeys: OptionKeys
+  optionKeys: OptionKeys<P>
   selectedOptions: string[]
   onChange?: (selectedOptions: string[]) => void
   onChangeCollapsed?: (value: boolean) => void
