+++ src/components/SearchSelect/types.ts
@@ -1,9 +1,10 @@
-export interface OptionKeys {
-  labelKey: string
-  valueKey: string
-  imageKey: string
+//TODO: Need remove string type
+export interface OptionKeys<P> {
+  labelKey: keyof P | string
+  valueKey: keyof P | string
+  imageKey: keyof P | string
   secondLabel?: {
-    key: string
+    key: keyof P | string
     format?: (value: number) => string
   }
 }
