+++ src/pages/LoansPage/components/LoansActiveTab/hooks/useLoansActiveTab.ts
@@ -2,6 +2,7 @@ import { useMemo, useState } from 'react'
 
 import { first, groupBy, map } from 'lodash'
 
+import { SearchSelectProps } from '@banx/components/SearchSelect'
 import { SortOption } from '@banx/components/SortDropdown'
 
 import { DEFAULT_SORT_OPTION } from '@banx/pages/LoansPage/constants'
@@ -12,38 +13,34 @@ import { useSortedLoans } from './useSortedLoans'
 
 import styles from '../LoansActiveTable.module.less'
 
-export type SearchSelectOption = {
+interface SearchSelectOption {
   collectionName: string
   collectionImage: string
+  numberOfNFTs: number
 }
 
 export const useLoansActiveTab = () => {
   const { loans, isLoading } = useWalletLoans()
 
   const [selectedOptions, setSelectedOptions] = useState<string[]>([])
-
   const [sortOption, setSortOption] = useState<SortOption>(DEFAULT_SORT_OPTION)
 
   const filteredLoans = useFilteredLoans(loans, selectedOptions)
   const sortedLoans = useSortedLoans(filteredLoans, sortOption.value)
 
   const searchSelectOptions = useMemo(() => {
-    const loansGroupedByCollection = groupBy(loans, (loan) => loan.nft.meta.collectionName)
+    const loansGroupedByCollection = groupBy(loans, ({ nft }) => nft.meta.collectionName)
 
     return map(loansGroupedByCollection, (groupedLoan) => {
       const firstLoanInGroup = first(groupedLoan)
       const { collectionName = '', collectionImage = '' } = firstLoanInGroup?.nft.meta || {}
       const numberOfNFTs = groupedLoan.length
 
-      return {
-        collectionName,
-        collectionImage,
-        numberOfNFTs,
-      }
+      return { collectionName, collectionImage, numberOfNFTs }
     })
   }, [loans])
 
-  const searchSelectParams = {
+  const searchSelectParams: SearchSelectProps<SearchSelectOption> = {
     onChange: setSelectedOptions,
     options: searchSelectOptions,
     selectedOptions,
@@ -57,12 +54,17 @@ export const useLoansActiveTab = () => {
     className: styles.searchSelect,
   }
 
+  const sortParams = {
+    option: sortOption,
+    onChange: setSortOption,
+  }
+
   return {
     loans: sortedLoans,
     loading: isLoading,
     sortViewParams: {
       searchSelectParams,
-      sortParams: { option: sortOption, onChange: setSortOption },
+      sortParams,
     },
   }
 }
