--- src/pages/OffersPage/components/PendingOfferTable/index.ts
@@ -1,1 +1,0 @@
-export * from './PendingOfferTable'
