+++ src/pages/OffersPage/components/PendingOffersTab/constants.ts
@@ -1,0 +1,6 @@
+import { SortOption } from '@banx/components/SortDropdown'
+
+export const DEFAULT_SORT_OPTION: SortOption = {
+  label: 'Apr',
+  value: 'apr_asc',
+}
