+++ src/pages/OffersPage/components/PendingOffersTable/PendingOffersTable.module.less
