+++ src/pages/OffersPage/components/PendingOffersTable/PendingOffersTable.tsx
