+++ src/pages/OffersPage/components/PendingOffersTable/TableCells/index.ts
