+++ src/pages/OffersPage/components/PendingOffersTable/columns.tsx
