+++ src/pages/OffersPage/components/PendingOffersTable/index.ts
@@ -1,0 +1,1 @@
+export * from './PendingOffersTable'
